Move PlanSummaryFormatter to proper location in source tree
