+++ arbeitszeit_flask/dependency_injection/presenters.py
@@ -15,6 +15,7 @@ from arbeitszeit_flask.url_index import CompanyUrlIndex, GeneralUrlIndex, Member
 from arbeitszeit_web.answer_company_work_invite import AnswerCompanyWorkInvitePresenter
 from arbeitszeit_web.colors import Colors
 from arbeitszeit_web.email import EmailConfiguration, MailService, UserAddressBook
+from arbeitszeit_web.formatters.plan_summary_formatter import PlanSummaryFormatter
 from arbeitszeit_web.get_company_summary import GetCompanySummarySuccessPresenter
 from arbeitszeit_web.get_company_transactions import GetCompanyTransactionsPresenter
 from arbeitszeit_web.get_coop_summary import GetCoopSummarySuccessPresenter
@@ -28,7 +29,6 @@ from arbeitszeit_web.language_service import LanguageService
 from arbeitszeit_web.list_all_cooperations import ListAllCooperationsPresenter
 from arbeitszeit_web.notification import Notifier
 from arbeitszeit_web.pay_means_of_production import PayMeansOfProductionPresenter
-from arbeitszeit_web.plan_summary_service import PlanSummaryFormatter
 from arbeitszeit_web.plotter import Plotter
 from arbeitszeit_web.presenters.accountant_invitation_presenter import (
     AccountantInvitationEmailPresenter,
