+++ arbeitszeit_web/get_plan_summary_company.py
@@ -2,8 +2,11 @@ from dataclasses import asdict, dataclass
 from typing import Any, Dict, Optional
 
 from arbeitszeit.use_cases.get_plan_summary_company import GetPlanSummaryCompany
+from arbeitszeit_web.formatters.plan_summary_formatter import (
+    PlanSummaryFormatter,
+    PlanSummaryWeb,
+)
 
-from .plan_summary_service import PlanSummaryFormatter, PlanSummaryWeb
 from .translator import Translator
 from .url_index import (
     EndCoopUrlIndex,
